Update VCDTypes.hpp
set VCDTime same as SignalTime i.e. long long

+++ include/vcd-parser/VCDTypes.hpp
@@ -20,7 +20,7 @@ typedef std::string VCDScopeName;
 typedef std::string VCDSignalHash;
 
 //! Represents a single instant in time in a trace
-typedef long VCDTime;
+typedef long long VCDTime;
 
 //! Specifies the timing resoloution along with VCDTimeUnit
 typedef long VCDTimeRes;
